Struct register::Field

pub struct Field<T, R> where
    T: IntLike,
    R: RegisterLongName,  {
    pub shift: usize,
    // some fields omitted
}

Specific section of a register.

Fields
shift: usize
Implementations

impl<T, R> Field<T, R> where
T: IntLike,
R: RegisterLongName,

pub fn read(self, val: T) -> T

pub fn is_set(self, val: T) -> bool

Check if one or more bits in a field are set

pub fn read_as_enum<E>(self, val: T) -> Option<E> where
E: TryFromValue<T, EnumType = E>,

Read value of the field as an enum member

impl<R> Field<u8, R> where
R: RegisterLongName,

pub const fn new(mask: u8, shift: usize) -> Field<u8, R>

pub fn val(&self, value: u8) -> FieldValue<u8, R>

impl<R> Field<u16, R> where
R: RegisterLongName,

pub const fn new(mask: u16, shift: usize) -> Field<u16, R>

pub fn val(&self, value: u16) -> FieldValue<u16, R>

impl<R> Field<u32, R> where
R: RegisterLongName,

pub const fn new(mask: u32, shift: usize) -> Field<u32, R>

pub fn val(&self, value: u32) -> FieldValue<u32, R>

impl<R> Field<u64, R> where
R: RegisterLongName,

pub const fn new(mask: u64, shift: usize) -> Field<u64, R>

pub fn val(&self, value: u64) -> FieldValue<u64, R>